Fixes for dots in filenames and <?php tags

+++ tests/rebuild.php
@@ -4,13 +4,17 @@
  * To rebuild the `snapshots` directory after changing
  * files in `source`, run `php tests/rebuild.php`.
  */
-exec('jigsaw build testing');
+shell_exec('./jigsaw build testing');
 removeDirectory('tests/snapshots');
 rename('tests/build-testing', 'tests/snapshots');
 
 function removeDirectory($path)
 {
-    $files = glob($path . '/*');
+    if (! $path) {
+        exit("Path to the 'tests/snapshots' directory is missing");
+    }
+
+    $files = glob($path . '/{,.}[!.,!..]*', GLOB_MARK|GLOB_BRACE);
 
     foreach ($files as $file) {
         is_dir($file) ? removeDirectory($file) : unlink($file);
